feat: add divs for Icons with w3.css
Each section first wraped in a overall class which is here is
w3-card-4. After that I added a img tag which will replace with Icons we
want in future. Next is text section which will be inside a container
and will be in the center of section. Inside that we will add out text
which here is "ICON" as template. These parts repeated 3 times.